react-leaflet v2 to v3 migration: Decouple WMSLayer component & some memos functions bodies
